fix(demo): make app-shell full-screen toggle demo-only
The expand-to-full-screen control was baked into the shell markup, so it showed
up in the code consumers copy. Move it into a docs-only wrapper
(application-shell-preview.tsx) shown as the variant's preview, while the shell
source (application-shell-demo.tsx) is now a clean, container-filling shell with
no full-screen button. The toggle sits in a slim top-right toolbar in the
wrapper (the shell's own top-right corner is occupied by its header controls).
